Arreglando navegación hacia la pantalla de dashboard
