+++ app/(auth)/_layout.tsx
@@ -9,7 +9,7 @@ export default function AuthLayout() {
 			<Stack.Screen name='login' options={{ headerShown: false }} />
 			<Stack.Screen name='register' options={{ title: 'Crear Cuenta' }} />
 			<Stack.Screen name='forgot-password' options={{ title: 'Recuperar Contraseña' }} />
-			<Stack.Screen name='dashboard' options={{ title: 'sesion iniciada' }} />
+
 			{/* Puedes agregar más pantallas aquí a medida que las crees */}
 		</Stack>
 	);
