+++ app/(tabs)/_layout.tsx
@@ -2,9 +2,9 @@ import { Tabs } from 'expo-router';
 import React from 'react';
 
 import { HapticTab } from '@/components/haptic-tab';
-import { IconSymbol } from '@/components/ui/icon-symbol';
 import { Colors } from '@/constants/theme';
 import { useColorScheme } from '@/hooks/use-color-scheme';
+import { Home } from 'lucide-react-native';
 
 export default function TabLayout() {
 	const colorScheme = useColorScheme();
@@ -17,21 +17,10 @@ export default function TabLayout() {
 				tabBarButton: HapticTab,
 			}}>
 			<Tabs.Screen
-				name='index'
+				name='dashboard'
 				options={{
-					title: 'Home',
-					tabBarIcon: ({ color }) => (
-						<IconSymbol size={28} name='house.fill' color={color} />
-					),
-				}}
-			/>
-			<Tabs.Screen
-				name='explore'
-				options={{
-					title: 'Explore',
-					tabBarIcon: ({ color }) => (
-						<IconSymbol size={28} name='paperplane.fill' color={color} />
-					),
+					title: 'Inicio',
+					tabBarIcon: ({ color }) => <Home color={color} size={28} strokeWidth={1.5} />,
 				}}
 			/>
 		</Tabs>
