+++ app/index.tsx
@@ -4,5 +4,9 @@ import { Redirect } from 'expo-router';
 export default function StartPage() {
 	const isAuthenticated = false;
 
-	return isAuthenticated ? <Redirect href='/(tabs)' /> : <Redirect href='/(auth)/home' />;
+	return isAuthenticated ? (
+		<Redirect href='/(tabs)/dashboard' />
+	) : (
+		<Redirect href='/(auth)/home' />
+	);
 }
