+++ package.json
@@ -47,7 +47,7 @@
 		"react-native-reanimated": "~4.1.1",
 		"react-native-safe-area-context": "~5.6.0",
 		"react-native-screens": "~4.16.0",
-		"react-native-svg": "^15.12.1",
+		"react-native-svg": "^15.13.0",
 		"react-native-web": "~0.21.0",
 		"react-native-worklets": "0.5.1",
 		"zod": "^3.25.76"
